Add back Nova hypervisor totals
Originally I had removed Nova hypervisor totals because a reviewer
suggested that the total calculation be deferred until query time.

This can be achieved by grouping by a particular period of time when
performing a sum aggregation, for example:

SELECT SUM(memory_mb) FROM openstack_hypervisor GROUP BY time(1m)

would sum all the hypervisor metrics within the same minute.

However, this is problematic because the query can't easily "know" what
the time interval between different data points is before the query is
executed. If data is sent every 30 seconds, and the query groups by a
time span of 1 minute, then the total in our example will show twice the
total hypervisor memory available in the environment.

This is especially problematic if you're trying to measure multiple
environments which may be sending metrics at different intervals.

Instead of dealing with all this at the query layer, this modifies the
plugin to send totals.