Update gophercloud dependency
This modifies our Godeps to pull in our custom gophercloud lib.

This is necessary because of a bug in upstream gophercloud about certain
assumptions about the hypervisors API. The bug is fixed in our fork.

Our gophercloud lib was built on top of the latest upstream master,
which made the small name switch in openstack.go necessary.